clean up examples
