+++ packages/styleguide/stories/Organisms/ConnectedForm/ConnectedFormGroup.stories.mdx
@@ -7,7 +7,7 @@ import LinkTo from '@storybook/addon-links/react';
 import {
   ConnectedFormPlayground,
   FormGroupStates,
-} from './ConnectedForm.examples.tsx';
+} from './ConnectedFormGroup.examples.tsx';
 
 <Meta
   title={title}
