+++ packages/styleguide/stories/Organisms/ConnectedForm/SubmitButton.stories.mdx
@@ -3,7 +3,7 @@ import title from '@codecademy/macros/lib/title.macro';
 import { Meta } from '@storybook/addon-docs/blocks';
 import LinkTo from '@storybook/addon-links/react';
 
-import { SubmitButtonStates } from './ConnectedForm.examples';
+import { SubmitButtonStates } from './SubmitButton.examples';
 
 <Meta
   title={title}
